+++ internal/modules/transactions/model/transaction_details.go
@@ -16,8 +16,8 @@ type TransactionDetail struct {
 	ProductID       int        `bun:"product_id,notnull"`
 	VariantID       *int       `bun:"variant_id,nullzero"` // Nullable
 	Quantity        int        `bun:"quantity,notnull"`
-	UnitPrice       int64      `bun:"unit_price,notnull"`
-	Subtotal        int64      `bun:"subtotal,notnull"`
+	UnitPrice       float64    `bun:"unit_price,notnull"`
+	Subtotal        float64    `bun:"subtotal,notnull"`
 	CreatedAt       time.Time  `bun:"created_at,notnull,default:current_timestamp"`
 	UpdatedAt       *time.Time `bun:"updated_at,nullzero"`
 	DeletedAt       *time.Time `bun:"deleted_at,soft_delete,nullzero"`
